% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_donnees.R
\name{preparer_base_senegal}
\alias{preparer_base_senegal}
\title{Préparer la base du Sénégal pour la cartographie}
\usage{
preparer_base_senegal(chemin_dta)
}
\arguments{
\item{chemin_dta}{Chemin du fichier Stata (.dta).}
}
\value{
Un dataframe avec deux colonnes : \code{NOMREG} (noms des régions) et
\code{QuantConsKGMoyIndAnnee} (quantité moyenne annuelle en kg).
}
\description{
Cette fonction lit un fichier \code{.dta}, agrège les données par ménage et calcule
la consommation moyenne annuelle par individu et par région.
}
\examples{
\dontrun{
base_senegal <- preparer_base_senegal("BASE_X1_Appuree.dta")
}
}
